third_party/blink/web_tests/external/wpt/app-history/navigate/navigate-no-args.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  await promise_rejects_js(t, TypeError, appHistory.navigate());
  await promise_rejects_js(t, TypeError, appHistory.navigate({}));
  await promise_rejects_js(t, TypeError, appHistory.navigate({ replace : true }));
}, "navigate() with no args or empty options");
</script>
